enh: Move ember-table query params init to mixin
