+++ app/controllers/events/list.js
@@ -1,14 +1,9 @@
 import Controller from '@ember/controller';
 import { computed, action } from '@ember/object';
+import EmberTableControllerMixin from 'open-event-frontend/mixins/ember-table-controller';
 
-export default class extends Controller {
-  queryParams = ['page', 'per_page'];
-  page = 1;
-  per_page = 10;
-  search = null;
-  sort_dir = null;
-  sort_by = null;
-  sorts = [];
+export default class extends Controller.extend(EmberTableControllerMixin) {
+
   @computed()
   get columns() {
     return [
